✨ Added running btn at HOME
