+++ src/components/util/base/ssBtn.vue
@@ -9,6 +9,7 @@ interface Prop {
   disable?: boolean
   loading?: boolean
   isCapital?: boolean
+  to?: string
   onClick: () => void
 }
 defineProps<Prop>()
@@ -24,6 +25,7 @@ defineProps<Prop>()
     :disable="loading || disable"
     :loading="loading"
     :no-caps="!isCapital"
+    :to="to"
     @click="onClick"
     :style="{ 'width': freeWidth ? '' : (width ?? '13rem') }"
   >
