+++ src/pages/WorldTabs/HomePage.vue
@@ -6,6 +6,7 @@ import BootSettingsView from 'src/components/World/HOME/OtherSettings/BootSettin
 import WorldDeleteView from 'src/components/World/HOME/WorldOperation/WorldDeleteView.vue';
 import WorldNameView from 'src/components/World/HOME/Top/WorldNameView.vue';
 import VersionSelecterView from 'src/components/World/HOME/Top/VersionSelecterView.vue';
+import RunningBtn from 'src/components/World/HOME/RunningBtn.vue';
 
 const scrollAreaRef = ref()
 
@@ -25,6 +26,8 @@ setScrollTop(scrollTop)
     style="flex: 1 1 0;"
   >
     <div class="mainField">
+      <RunningBtn to="/console" :test-font-size="1.1" class="full-width q-mt-xl" />
+
       <div class="row justify-center q-mt-sm q-pb-lg q-gutter-lg">
         <div class="col" style="min-width: 12rem;">
           <h1 class="q-pt-none">{{ $t("home.worldName.title") }}</h1>
